front_container_with_model/front_container_with_model/requirements.txt
flask
tensorflow
numpy
os
werkzeug.utils
requests
Pillow